Debug: Move DMI-specific code to a seperate file, to reduce size of Debug.scala
